# Optimize batch index ACK performance